Moved Class to Separate File
Moved the contents by using the shortcut {CTRL} {DOT} to move the class to a new file named Person.cs.  You can see this in the Solutions Eplorer.  If Solutions Eplorer is not Open {CTRL} + ALT + L.

Note to Open the File you can also click in the word "Person" where it says new Person and press {F12}.  This will open the class in a popout window.

To keep the Window Open {CTRL} + {ALT} + {HOME}.  When you do this the Person Tab will move from the top right to the top left.