Moved Class
In this step I moved the Class to the CSLib Project using the shortcut {SHIFT} + {LEFT_CLICK} and drag to the CSLIB project.

Note:  This closed the Person.cs file that was opened because the file was no longer available that the tab was pointing to.

To open you can use the same trick.  By clicking in the name of the class and pressing {F12}.  To Keep open permanently {{CTRL} + {ALT} + {HOME}.

Note:  If you start to type anything your popup window for Person.cs will close automatically.