Wrapping Code Up
Visual studio allows us to use pre compilation commands.  They start with a # symbol.  For this section I have created a region named Introduction to Classes.

This is a way of organizing related methods together.

Note:  The difference between a Function and a method is that a method does not return a value.  To most programmers Methods and Functions are used Interchangeably.