New Field for List of Orders
This Type List requires a using statement to be added.  the Using statement can be added by using {CTRL} + {ENTER} to open the Class View and select the option to add the appropriate using object.

The Class View is helpful in many ways and can be navigated using the {UP} and {DOWN} arrows.