Hierarchy in Constructors
Here we can use the base class constructor to define a hierarchy and prevent duplication of code.

If coding this we would use the object initializer for the id and name fields and only need the default constructor ().  However for this case I wanted to show clearly how class constructors can be used.