Readonly Modifiier
Here in the Customers.cs we set the field List of Orders as readonly.  After Building you will receive an error.  "A readonly field cannot be assigned to (except in a constructor or a variable initializer).

Note:  If you double click it will take you to the promote method.  And you will need to remove this code.