New Instance of Customer Class
We will start here be creating a new instance of the customer class.  Now on the next line you can see when accessing your customer class that you have the Promote method which we created.  But we have other methods like ToString(), GetType(), GetHash() and several others.  Those represent methods that are derrived from the Base Class of .NET Framework in which all classes are derrived from.