Boxing and Unboxing
Here we show that primitive data types like int, bool, float, long, struct, etc... are placed on the stack and objects are placed on the heap.  We box objects to the heap and unbox from the heap back to the stack.

By using an ArrayList() boxing and unboxing occurs.  The pointer used for the object from the stack to the heap causes overhead.  The overhead on large ArrayLists can cause performance issues.  By delcaring a generic list with a data type we have all the functionality of an ArrayList and place the work on the stack with none of the overhead required.  A limitation here is  only one data type can be used however in an ArrayList many different data types can be used and you will need to use type casting to access the data since all the data is stored as an object in the stack with a pointer to the heap.