Overriding the Draw Method
Now when adding a shape to a List of Shapes we can add a Circle, Rectange, or a Triangle to the list.  Then when we call the method to DrawShapes in the Canvas class. we can itterate through the list of shapes and draw all three implementations on the screen.