Enforce Override
By declaring a method and a class abstrct it forces developers to implement you code in a specific way.

In the default example ShapeText will throw an error before you add the override for the abstract Draw method.

Essentially by decalaring a method abstract you do not implement the method in the base class.  Insteady you must override the method in all Derrived Classes.

This forces developers to use the structure that you have impleemtned.