Class Abstract Why
Here I try to create a new intance of the base class.  However when building the Shape I receive an error.

Cannot create an instance of the abstract class or interface 'Shape'

This is caused becassue the class is abstract and this is a good thing.  Here we can imagine a shape as polygon with no real form.  So this would prevent an end user from